+++ src/main/java/com/vention/agroex/model/BetRequest.java
@@ -2,8 +2,9 @@ package com.vention.agroex.model;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
 @Builder
-public record BetRequest(UUID userId, Float amount, Instant betTime) {}
+public record BetRequest(UUID userId, BigDecimal amount, Instant betTime) {}
