+++ src/main/java/com/vention/agroex/model/LotStatusResponse.java
@@ -2,9 +2,10 @@ package com.vention.agroex.model;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.UUID;
 
 @Builder
-public record LotStatusResponse(Long lotId, UUID userid, Float betAmount, Instant betTime, String status) {
+public record LotStatusResponse(Long lotId, UUID userid, BigDecimal betAmount, Instant betTime, String status) {
 }
